requirements.txt
fastapi 
googlemaps 
open-webui 
python-dotenv 
slowapi
uvicorn 